requirements.txt
django-oscar[sorl-thumbnail]==3.2.2
pycountry==23.12.11
wagtail>=4.1,<4.2
python-dotenv==1.0.1